refactor(linux): Move GSettings code to separate class
